Continuing with API changes.

+++ SimpleSchedulerModels/Worker.cs
@@ -1,5 +1,5 @@
 namespace SimpleSchedulerModels;
 
 public record class Worker(long ID, bool IsActive, string WorkerName, string DetailedDescription,
-    string EmailOnSuccess, long? ParentWorkerID, long TimeoutMinutes,
+    string EmailOnSuccess, long? ParentWorkerID, int TimeoutMinutes,
     string DirectoryName, string Executable, string ArgumentValues);
