+++ SqlDB/app/Types/WorkersType.sql
@@ -5,7 +5,7 @@
     ,[DetailedDescription] NVARCHAR(MAX) NOT NULL
     ,[EmailOnSuccess] NVARCHAR(MAX) NOT NULL
     ,[ParentWorkerID] BIGINT NULL
-    ,[TimeoutMinutes] BIGINT NOT NULL
+    ,[TimeoutMinutes] INT NOT NULL
     ,[DirectoryName] NVARCHAR(1000) NOT NULL
     ,[Executable] NVARCHAR(1000) NOT NULL
     ,[ArgumentValues] NVARCHAR(1000) NOT NULL
